refactor: destructurization FormContent component
