+++ ui/form/src/index.ts
@@ -1,1 +1,2 @@
 export * from './form.component'
+export * from './privacy-policy'
