+++ ui/form/src/privacy-policy/index.ts
@@ -1,0 +1,1 @@
+export * from './privacy-policy.component'
